restructure code and remove unused ones
